Split bulk SMS sending functionality into a new CommunicationsManager
