+++ src/CareTogether.Core/Managers/Records/RecordsManager.cs
@@ -2,9 +2,7 @@
 using CareTogether.Resources.Approvals;
 using CareTogether.Resources.Directory;
 using CareTogether.Resources.Notes;
-using CareTogether.Resources.Policies;
 using CareTogether.Resources.Referrals;
-using CareTogether.Utilities.Telephony;
 using Nito.AsyncEx;
 using System;
 using System.Collections.Generic;
@@ -22,22 +20,18 @@ namespace CareTogether.Managers.Records
         private readonly IApprovalsResource approvalsResource;
         private readonly IReferralsResource referralsResource;
         private readonly INotesResource notesResource;
-        private readonly IPoliciesResource policiesResource;
-        private readonly ITelephony telephony;
         private readonly CombinedFamilyInfoFormatter combinedFamilyInfoFormatter;
 
 
         public RecordsManager(IAuthorizationEngine authorizationEngine, IDirectoryResource directoryResource,
             IApprovalsResource approvalsResource, IReferralsResource referralsResource, INotesResource notesResource,
-            IPoliciesResource policiesResource, ITelephony telephony, CombinedFamilyInfoFormatter combinedFamilyInfoFormatter)
+            CombinedFamilyInfoFormatter combinedFamilyInfoFormatter)
         {
             this.authorizationEngine = authorizationEngine;
             this.directoryResource = directoryResource;
             this.approvalsResource = approvalsResource;
             this.referralsResource = referralsResource;
             this.notesResource = notesResource;
-            this.policiesResource = policiesResource;
-            this.telephony = telephony;
             this.combinedFamilyInfoFormatter = combinedFamilyInfoFormatter;
         }
 
@@ -164,54 +158,6 @@ namespace CareTogether.Managers.Records
                     throw new NotImplementedException(
                         $"The command type '{command.GetType().FullName}' has not been implemented.");
             }
-        }
-
-        public async Task<ImmutableList<(Guid FamilyId, SmsMessageResult? Result)>> SendSmsToFamilyPrimaryContactsAsync(
-            Guid organizationId, Guid locationId, ClaimsPrincipal user,
-            ImmutableList<Guid> familyIds, string sourceNumber, string message)
-        {
-            if (!await authorizationEngine.AuthorizeSendSmsAsync(organizationId, locationId, user))
-                throw new Exception("The user is not authorized to perform this command.");
-
-            // Validate that the requested source number has been configured for the specified location.
-            var configuration = await policiesResource.GetConfigurationAsync(organizationId);
-            var sourcePhoneNumber = configuration.Locations
-                .Single(location => location.Id == locationId)
-                .SmsSourcePhoneNumbers.SingleOrDefault(number => number.SourcePhoneNumber == sourceNumber)
-                ?.SourcePhoneNumber;
-
-            if (sourcePhoneNumber == null)
-                throw new InvalidOperationException(
-                    "The specified location does not have the requested source phone number configured for SMS messages.");
-
-            var families = await directoryResource.ListFamiliesAsync(organizationId, locationId);
-
-            var familyPrimaryContactNumbers = familyIds
-                .Select(familyId =>
-                {
-                    var family = families.Single(family => family.Id == familyId);
-                    var primaryContactAdult = family.Adults
-                        .Single(adult => adult.Item1.Id == family.PrimaryFamilyContactPersonId);
-                    return (familyId,
-                        phoneNumber: primaryContactAdult.Item1.PhoneNumbers
-                            .SingleOrDefault(number => number.Id == primaryContactAdult.Item1.PreferredPhoneNumberId));
-                }).ToImmutableList();
-
-            var destinationNumbers = familyPrimaryContactNumbers
-                .Where(x => x.phoneNumber != null)
-                .Select(x => x.phoneNumber!.Number)
-                .Distinct()
-                .ToImmutableList();
-
-            var sendResults = await telephony.SendSmsMessageAsync(sourcePhoneNumber, destinationNumbers, message);
-
-            var allFamilyResults = familyPrimaryContactNumbers.Select(x =>
-            {
-                var sendResult = sendResults.SingleOrDefault(result => result.PhoneNumber == x.phoneNumber?.Number);
-                return (FamilyId: x.familyId, Result: sendResult);
-            }).ToImmutableList();
-
-            return allFamilyResults;
         }
 
         public async Task<CombinedFamilyInfo> ExecuteAtomicRecordsCommandAsync(Guid organizationId, Guid locationId,
